+++ layers/GatoGraphQLForWP/phpunit-packages/gatographql/tests/Integration/DownToEditorSchemaEditingAccessModifyPluginSettingsFixtureEndpointWebserverRequestTestTrait.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PHPUnitForGatoGraphQL\GatoGraphQL\Integration;
+
+use GatoGraphQL\GatoGraphQL\Security\UserAuthorizationSchemes\DownToEditorRoleUserAuthorizationScheme;
+use GatoGraphQL\GatoGraphQL\Security\UserAuthorizationSchemes\UserAuthorizationSchemeInterface;
+
+trait DownToEditorSchemaEditingAccessModifyPluginSettingsFixtureEndpointWebserverRequestTestTrait
+{
+    protected function getUserAuthorizationScheme(): UserAuthorizationSchemeInterface
+    {
+        return new DownToEditorRoleUserAuthorizationScheme();
+    }
+}